Fix return types

+++ audmetric/core/api.py
@@ -656,7 +656,7 @@ def linkability(
     # Global measure using trapz numerical integration
     d_sys = np.trapz(x=bin_centers, y=d * y1)
 
-    return d_sys
+    return float(d_sys)
 
 
 def mean_absolute_error(
@@ -997,7 +997,7 @@ def unweighted_average_bias(
     if denominator == 0:
         return np.nan
 
-    return bias / denominator
+    return float(bias / denominator)
 
 
 def unweighted_average_fscore(
@@ -1225,7 +1225,7 @@ def word_error_rate(
         wer += edit_distance(t, p) / n
 
     num_samples = len(truth) if len(truth) > 1 else 1
-    return wer / num_samples
+    return float(wer / num_samples)
 
 
 def _matching_scores(
